feat(UI): Imrpoved results page with grid-based layout. Completed 'bias result' card, and have set out structure for the analyzed text and keywords cards.
