+++ src/app/results/components/AnalysedText.tsx
@@ -1,0 +1,7 @@
+"use client";
+
+import { Typography } from "@mui/material";
+
+export function AnalysedText() {
+  return <Typography>AnalyzedText</Typography>;
+}
